% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_configuration}
\alias{get_bo_configuration}
\title{Getting the Configuration parameters}
\usage{
get_bo_configuration(domain, type = "functions")
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{type}{This parameter is obligatory. By default set to \code{"functions"}.
Indicates the type from which to be retrieved. Other possible values are:
\itemize{
\item \code{"functions"}
\item \code{"charsets"}
\item \code{"visualizations"}
\item \code{"palettes"}
\item \code{"formats"}
\item \code{"fontmappings"}
\item \code{"operators"}
\item \code{"skins"}
}}
}
\value{
Returns a tidy \code{data.frame} of SAP Business Objects configuration details
}
\description{
Retrieves information on default configuration details declared on the WACS server, such as the default formats or color palettes.
}
\examples{
\dontrun{
get_bo_configuration(domain="YOUR_DOMAIN", type="charsets")
get_bo_configuration(domain="YOUR_DOMAIN", type="functions")
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/0c4ef646601e4c1eaf9a2571091e33d5.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
