% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_connections}
\alias{get_bo_connections}
\title{Retrieving the Connections}
\usage{
get_bo_connections(domain, type = NULL, offset = 0, limit = 50)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{type}{Indicates the connection type to be retrieved. Values are:
\itemize{
\item \code{"Relational"}
\item \code{"FlattenedOlap"}
\item \code{"Olap"}
\item \code{"DataFederator"}
}
This parameter is optional. If not specified, all types of connections are retrieved.}

\item{offset}{Indicates the position in the list, from which connections are returned.
It must be greater than or equal to 0. The default value is  \code{0}. This parameter is optional.}

\item{limit}{indicates the number of connections that you can list on one page. Its range is 1 to 50.
This parameter is optional. The default value is \code{50}.}
}
\value{
Returns a tidy \code{data.frame} of SAP Business Objects connections
}
\description{
Retrieves the list of SAP Business Objects connections stored in the CMS repository that the end-user has access to.
}
\examples{
\dontrun{
get_bo_connections(domain="YOUR_DOMAIN")
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/cc73a43eae76475291d35e1c3ef5451d.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
