% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_document_details}
\alias{get_bo_document_details}
\title{Retrieve Business Objects document details}
\usage{
get_bo_document_details(domain, document_id, tracker_document_id = NULL)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{document_id}{Document ID of the SAP Business Objects Document}

\item{tracker_document_id}{Optional parameter. Identifier of a reference document for trackdata feature.
Must be provided only when the document state is unused.}
}
\value{
Returns a tidy \code{data.frame} with detailed information about the document.
}
\description{
Retrieves Gets the details of a Web Intelligence document and returns a tidy \code{data.frame} with detailed information.
}
\examples{
\dontrun{
get_bo_document_details(domain="YOUR_DOMAIN",
                       document_id="YOUR_DOCUMENT_ID")
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/ec5627ba6fdb101497906a7cb0e91070.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
