% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_document_schedules}
\alias{get_bo_document_schedules}
\title{Retrieve Business Objects document schedules}
\usage{
get_bo_document_schedules(domain, document_id)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{document_id}{Document ID of the SAP Business Objects Document}
}
\value{
Returns a tidy \code{data.frame} with detailed information about the documents' schedules.
}
\description{
Retrieves the schedules for a specific document ID and returns a tidy \code{data.frame} with detailed information.
}
\examples{
\dontrun{
get_bo_document_schedules(domain="YOUR_DOMAIN",
                          document_id="YOUR_DOCUMENT_ID")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
