% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_folder_ressources}
\alias{get_bo_folder_ressources}
\title{Retrieve all the children elements of a folder}
\usage{
get_bo_folder_ressources(
  domain,
  folder_id,
  type = NULL,
  page = NULL,
  pageSize = 50
)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{folder_id}{Folder ID in the CMS repository of the folder to retrieve its children from (Equals to the \code{SI_ID} variable).}

\item{type}{Optional parameter to define the type of elements to retrieve. Must be a single element (Equals to the \code{SI_KIND} variable).
Some of the types of the elements to retrieve are the following (incomplete list):
\itemize{
\item \code{"InfoView"}
\item \code{"Webi"}
\item \code{"Publication"}
\item \code{"Folder"}
}}

\item{page}{For more than 50 elements within a folder pagination is necessary. Use this parameter to control pagination.}

\item{pageSize}{Page size of the number of ressources to obtain. Maximum is 50, defaults to 50.}
}
\value{
Returns a tidy \code{data.frame} of all the children elements within the folder
}
\description{
Retrieves all the children ressources with name, description, id and type from a given Folder ID.
The current file and the one to upload must have the same file structure (columns number, names, and order).
}
\examples{
\dontrun{
get_bo_folder_ressources(domain = "YOUR_DOMAIN",
                         folder_id = 12345,
                         type = Webi
)
}
}
\references{
\url{https://help.sap.com/viewer/db6a17c0d1214fd6971de66ea0122378/4.2.3/en-US/45aa8bae6e041014910aba7db0e91070.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
