% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_universe_details}
\alias{get_bo_universe_details}
\title{Getting the details of a SAP Business Objects' universe}
\usage{
get_bo_universe_details(domain, universe_id, aggregated = FALSE)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{universe_id}{Universe ID from which to obtain the details}

\item{aggregated}{Is an optional, boolean/logical parameter that indicates if the outline must be aggregated. Default value is \code{FALSE}.}
}
\value{
Returns the SAP Business Objects Universes details
}
\description{
Gets the details of a UNX or UNV universe referenced by its ID.
}
\examples{
\dontrun{
get_bo_universe_details(domain="YOUR_DOMAIN",
                        universe_id=12345)
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/ec5584596fdb101497906a7cb0e91070.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
