% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bo_universes}
\alias{get_bo_universes}
\title{Getting the List of SAP Business Objects universes}
\usage{
get_bo_universes(domain, offset = 0, limit = 50)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{offset}{Indicates the position in the list, from which universes are returned.
It must be greater than or equal to 0. The default value is 0. This parameter is optional. The default value is \code{0}.}

\item{limit}{Indicates the number of universes that you can list on one page. The range is from 1 to 50.
The default value is 10. This parameter is optional. The default value is \code{50}.}
}
\value{
Returns a tidy \code{data.frame} of SAP Business Objects Universes
}
\description{
Gets the list of UNX and UNV universes stored in the CMS repository.
}
\examples{
\dontrun{
get_bo_universes(domain="YOUR_DOMAIN")
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/ec558f026fdb101497906a7cb0e91070.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
