% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_schedule_details}
\alias{get_schedule_details}
\title{Retrieve Business Objects document' schedule details}
\usage{
get_schedule_details(domain, document_id, schedule_id)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{document_id}{Document ID of the SAP Business Objects Document}

\item{schedule_id}{Schedule ID of the specific SAP Business Objects Document. Can be obtained through the \code{get_document_schedules()} function.}
}
\value{
Returns details about the SAP Business Objects schedule.
}
\description{
Retrieves the details of a specific schedule ID of a SAP Business Objects document with information about document parameters & frequency.
}
\examples{
\dontrun{
get_schedule_details(domain="YOUR_DOMAIN",
                     document_id="YOUR_DOCUMENT_ID",
                     schedule_id="YOUR_SCHEDULE_ID")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
