% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_off}
\alias{log_off}
\title{Logs off from the SAP Business Objects}
\usage{
log_off(domain)
}
\arguments{
\item{domain}{SAP Business Objects domain to log off from}
}
\value{
Does not return any object, only a success message, as SAP API token is removed from the environment.
}
\description{
Logs off from the SAP Business Objects session and removes the SAP Business Objects API token from the System Environment
}
\section{Warning}{
 Please be careful when dealing with API keys and other
secrets & tokens - keep them private and do not publish them.
}

\examples{
\dontrun{
log_off(domain="YOUR_DOMAIN")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
