% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_on}
\alias{log_on}
\title{Logs into SAP Business Objects}
\usage{
log_on(domain, username, password, authentication_type = "secLDAP")
}
\arguments{
\item{domain}{SAP Business Objects domain to log into}

\item{username}{SAP Business Objects username}

\item{password}{SAP Business Objects password}

\item{authentication_type}{Authentication Type to logon into SAP Business Objects. Defaults to \code{secLDAP}. Must be one of the following:
\itemize{
\item secLDAP
\item secEnterprise
\item secWinAD
\item secSAPR3
}}
}
\value{
Does not return any object, only a success message, as SAP API token is saved into the environment to work with further functions.
}
\description{
Logs into SAP Business Objects and saves the SAP Business Objects API token into the System Environment
}
\section{Warning}{
 Please be careful when dealing with API keys and other
secrets & tokens - keep them private and do not publish them.
}

\examples{
\dontrun{
log_on(domain="YOUR_DOMAIN",
       username="YOUR_USERNAME",
       password="YOUR_PASSWORD",
       authentication_type="secLDAP")
}
}
\author{
Matthias Brenninkmeijer -
\href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
