% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_bo_excel_file}
\alias{update_bo_excel_file}
\title{Update an Excel File from SAP Business Objects}
\usage{
update_bo_excel_file(domain, excel_file_path, spreadsheet_id)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{excel_file_path}{Location of the Excel file. Must include directory, file name and file extension.
Must be a Microsoft Excel 2003 or Microsoft Excel 2007 format file.}

\item{spreadsheet_id}{ID of the Excel file in the CMS repository.}
}
\value{
Returns a message stating the success or failure of the request.
}
\description{
Updates an already-stored Microsoft Excel file in the CMS repository.
The current file and the one to upload must have the same file structure (columns number, names, and order).
}
\examples{
\dontrun{
update_bo_excel_file(domain = "YOUR_DOMAIN",
                   excel_file_path = "/Users/YOurUsername/Downloads/example_file.xlsx",
                   spreadsheet_id = 123456
)
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/cb615e388c784ef193c1873455e7ae39.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
