% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_excel_to_bo}
\alias{upload_excel_to_bo}
\title{Upload an Excel File to SAP Business Objects}
\usage{
upload_excel_to_bo(domain, excel_file_path, folder_id)
}
\arguments{
\item{domain}{SAP Business Objects domain}

\item{excel_file_path}{Location of the Excel file. Must include directory, file name and file extension.
Must be a Microsoft Excel 2003 or Microsoft Excel 2007 format file.}

\item{folder_id}{SAP Business Objects folder ID to which the Microsoft Excel file should be stored.}
}
\value{
Returns a message stating the success or failure of the request.
}
\description{
Uploads and stores a Microsoft Excel file to the CMS repository.
}
\examples{
\dontrun{
upload_excel_to_bo(domain = "YOUR_DOMAIN",
                   excel_file_path = "/Users/YOurUsername/Downloads/example_file.xlsx",
                   folder_id = 123456
)
}
}
\references{
\url{https://help.sap.com/viewer/58f583a7643e48cf944cf554eb961f5b/4.2/en-US/cb615e388c784ef193c1873455e7ae39.html}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
